check for distributed

+++ one-shot-tuning.py
@@ -353,11 +353,11 @@ def main(
             print(f"max memory allocated: {max_memory_allocated:.3f} GB.")
             with accelerator.accumulate(unet):
                 # Convert videos to latent space
-                pixel_values = batch["pixel_values"].to(weight_dtype).half()
+                pixel_values = batch["pixel_values"].to(weight_dtype)
                 video_length = pixel_values.shape[1]
                 pixel_values = rearrange(pixel_values, "b f c h w -> (b f) c h w")
                 check_gpu()
-                latents = vae.encode(pixel_values).latent_dist.sample()
+                latents = vae.encode(pixel_values.half()).latent_dist.sample()
                 latents = rearrange(latents, "(b f) c h w -> b c f h w", f=video_length)
                 latents = latents * 0.18215
